Bugly修改项目，位置时间戳改为long,百度地图修改初始化位置,加锁防数组越界，添加 位置定位时间

+++ app/src/main/java/com/xiaomaoqiu/now/bean/nocommon/PetLocationBean.java
@@ -9,7 +9,7 @@ public class PetLocationBean extends BaseBean{
 
     public double latitude;
 
-    public double location_time;
+    public long location_time;
 
     public double longitude;
